backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
psycopg[binary]>=3.1.0
python-dotenv>=1.0.0
pydantic>=2.9.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0
python-multipart>=0.0.9
anthropic>=0.42.0
google-genai>=1.0.0
PyMuPDF>=1.24.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
httpx>=0.27.0
boto3
